api/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
rembg>=2.0.56
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.25.0
onnxruntime
python-multipart

# Phase 1: Core AI Infrastructure
torch>=2.0.0
torchvision>=0.15.0
segment-anything>=1.0
ultralytics>=8.0.0
transformers>=4.30.0

# Phase 2: Image Enhancement
# Note: These packages require Python 3.11 or earlier due to build compatibility
# Uncomment if using Python 3.11 or earlier:
# basicsr>=1.4.2
# realesrgan>=0.3.0
# gfpgan>=1.3.8

# Alternative for Python 3.13: Use these lightweight alternatives
opencv-contrib-python>=4.8.0
scikit-image>=0.22.0

# Phase 3: Background Generation (prepared for later)
diffusers>=0.21.0
accelerate>=0.20.0

# Utilities
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
